Move builtins compat to separate file.
